Exception tweaks.
In the event of an exception, the rackspace libraries try to match
information to an exception.  If the exception does not match a defined
exception, but does match an existing class name, say, an active record
model, then we get unexpected and undesireable behaviour.  This attempts
to resolve that issue.